# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_AVGsync_analysis-pro_backend_internal_model_request.LoginRequest:
    properties:
      email:
        description: Email зарегистрированного пользователя.
        example: ivan.petrov@example.com
        type: string
      password:
        description: Пароль пользователя.
        example: Str0ngPass!2026
        type: string
    required:
    - email
    - password
    type: object
  github_com_AVGsync_analysis-pro_backend_internal_model_request.NewUserRequest:
    properties:
      email:
        description: Уникальный email, используется для входа.
        example: ivan.petrov@example.com
        type: string
      full_name:
        description: Полное имя пользователя, отображается в профиле.
        example: Иван Петров
        maxLength: 64
        minLength: 3
        type: string
      password:
        description: Пароль, минимум 8 символов.
        example: Str0ngPass!2026
        minLength: 8
        type: string
    required:
    - email
    - full_name
    - password
    type: object
  github_com_AVGsync_analysis-pro_backend_internal_model_request.UserUpdateRequest:
    properties:
      email:
        description: Новый email. Если поле не передано, текущее значение не меняется.
        example: ivan.new@example.com
        type: string
      full_name:
        description: Новое полное имя. Если поле не передано, текущее значение не
          меняется.
        example: Иван Сергеевич Петров
        maxLength: 64
        minLength: 3
        type: string
    type: object
  github_com_AVGsync_analysis-pro_backend_internal_model_response.UserResponse:
    properties:
      email:
        example: ivan.petrov@example.com
        type: string
      full_name:
        example: Иван Петров
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      role:
        enum:
        - user
        - admin
        example: user
        type: string
      subscription_expires:
        example: "2026-12-31T23:59:59Z"
        type: string
      subscription_plan:
        enum:
        - free
        - pro
        - enterprise
        example: free
        type: string
    type: object
info:
  contact: {}
  description: |-
    REST API для Analysis Pro: авторизация, профиль пользователя, детализация продаж, прогнозы по товарам, рекомендации по закупкам и экспорт CSV/XML.

    Схема авторизации: вызовите `POST /auth/login`; API установит HttpOnly cookie `token`. В DEBUG режиме также можно передавать `Authorization: Bearer <jwt>`.
    Swagger UI доступен только при `DEBUG=true`: `/swagger`, JSON спецификация — `/swagger/doc.json`.
  title: API Analysis Pro
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: |-
        Проверяет email и пароль пользователя.
        При успехе устанавливает JWT в HttpOnly cookie с именем `token`.
        Защищённые маршруты используют cookie `token`; в DEBUG режиме также принимается заголовок `Authorization: Bearer <jwt>`.
      parameters:
      - description: 'Данные входа: email и пароль. Примеры значений указаны в схеме.'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AVGsync_analysis-pro_backend_internal_model_request.LoginRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          headers:
            Set-Cookie:
              description: token=<jwt>; HttpOnly; Secure; SameSite=Strict; Path=/;
                Max-Age=86400
              type: string
          schema:
            type: string
        "400":
          description: Некорректное тело запроса
          schema:
            type: string
        "401":
          description: Неверный email или пароль
          schema:
            type: string
      summary: Войти в аккаунт
      tags:
      - Авторизация
  /auth/register:
    post:
      consumes:
      - application/json
      description: |-
        Создаёт аккаунт по полному имени, email и паролю.
        Email должен быть уникальным. Пароль хранится в виде хеша и никогда не возвращается в ответе.
        При успехе возвращает профиль созданного пользователя без токена; после регистрации вызовите /auth/login.
      parameters:
      - description: 'Данные регистрации: полное имя, email и пароль. Примеры значений
          указаны в схеме.'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AVGsync_analysis-pro_backend_internal_model_request.NewUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Профиль созданного пользователя
          schema:
            $ref: '#/definitions/github_com_AVGsync_analysis-pro_backend_internal_model_response.UserResponse'
        "400":
          description: Некорректное тело запроса
          schema:
            type: string
        "500":
          description: Не удалось зарегистрировать пользователя
          schema:
            type: string
      summary: Зарегистрировать нового пользователя
      tags:
      - Авторизация
  /me:
    get:
      description: |-
        Возвращает профиль пользователя, определённого по JWT claims.
        Требует валидную cookie `token`, полученную через /auth/login. В DEBUG режиме также принимается `Authorization: Bearer <jwt>`.
      produces:
      - application/json
      responses:
        "200":
          description: Профиль текущего пользователя
          schema:
            $ref: '#/definitions/github_com_AVGsync_analysis-pro_backend_internal_model_response.UserResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            type: string
        "500":
          description: Не удалось получить пользователя
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить профиль текущего пользователя
      tags:
      - Пользователи
    patch:
      consumes:
      - application/json
      description: |-
        Обновляет email и/или полное имя текущего пользователя.
        Все поля необязательные. Если поле не передано, прежнее значение сохраняется.
        Требует валидную cookie `token`, полученную через /auth/login. В DEBUG режиме также принимается `Authorization: Bearer <jwt>`.
      parameters:
      - description: 'Данные обновления профиля: email и/или полное имя. Примеры значений
          указаны в схеме.'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AVGsync_analysis-pro_backend_internal_model_request.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Обновлённый профиль пользователя
          schema:
            $ref: '#/definitions/github_com_AVGsync_analysis-pro_backend_internal_model_response.UserResponse'
        "400":
          description: Некорректное тело запроса
          schema:
            type: string
        "401":
          description: Пользователь не авторизован
          schema:
            type: string
        "500":
          description: Не удалось обновить пользователя
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновить профиль текущего пользователя
      tags:
      - Пользователи
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Только для DEBUG режима. Введите: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"